Rename posocc.h -> hit.h.
